philosophers

+++ inc/philo.h
@@ -6,7 +6,7 @@
 /*   By: lraffin <lraffin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/15 14:27:28 by lraffin           #+#    #+#             */
-/*   Updated: 2021/12/23 01:45:11 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/12/23 23:47:29 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,33 +28,33 @@
 # define DIED "\e[91mdied\e[0m"
 
 /*** init ***/
-t_data	*init_data(int ac, char **av);
-t_bool	create_philos(t_data *data);
-void	join_philos(t_data *data);
+t_data		*init_data(int ac, char **av);
+t_bool		create_philos(t_data *data);
+void		join_philos(t_data *data);
 
 /*** parsing ***/
-t_bool	check_input(int ac, char **av);
+t_bool		check_input(int ac, char **av);
 
 /*** routine ***/
-void	*routine(void *arg);
-void	eat_action(t_philo *philo, int i);
+void		*routine(void *arg);
+void		eat_action(t_philo *philo, int i);
 
 /*** exit ***/
-void	exit_error(char *error_msg, int exit_code);
-void	exit_error_free(char *error_msg, int exit_code, t_data *data);
-void	clean_data(t_data *data);
+void		exit_error(char *error_msg, int exit_code);
+void		exit_error_free(char *error_msg, int exit_code, t_data *data);
+void		clean_data(t_data *data);
 
 /*** utils ***/
-void	ft_putstr_fd(int fd, char *s);
-long	ft_atoi(char *s);
+void		ft_putstr_fd(int fd, char *s);
+long long	ft_atoi(char *s);
 /* mutex */
-void	write_mutex(t_mutex *mutex, int value);
-int		read_mutex(t_mutex *mutex);
+void		write_mutex(t_mutex *mutex, int value);
+int			read_mutex(t_mutex *mutex);
 /* time*/
-int		get_time(void);
-void	ft_usleep(int time);
+int			get_time(void);
+void		ft_usleep(int time);
 /* print */
-void	update_status(char *msg, t_philo *philo, int death);
-void	prompt(void);
+void		update_status(char *msg, t_philo *philo, int death);
+void		prompt(void);
 
 #endif
