+++ src/utils/utils.c
@@ -6,7 +6,7 @@
 /*   By: lraffin <lraffin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/16 01:51:15 by lraffin           #+#    #+#             */
-/*   Updated: 2021/12/19 02:13:05 by lraffin          ###   ########.fr       */
+/*   Updated: 2021/12/23 23:48:03 by lraffin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,11 +18,11 @@ void	ft_putstr_fd(int fd, char *s)
 		write(fd, s++, 1);
 }
 
-long	ft_atoi(char *s)
+long long	ft_atoi(char *s)
 {
-	unsigned long long	value;
-	int					sign;
-	size_t				i;
+	long long	value;
+	long long	sign;
+	int			i;
 
 	value = 0;
 	sign = 1;
